refactor(frontend): move the React out of lib/
Four files under `src/lib` were not library code:

- `projects.tsx` (523 lines) and `settings.tsx` (301) are context Providers —
  the app's state layer, with 13 and 9 importers between them. They move to
  `src/providers/`, and `settings-clamp.test.ts` follows what it tests.
- `file-icon.tsx` and `provider-icons.tsx` each export exactly one component.
  They move to `src/components/` under the name they export, the convention
  every other component there already follows: `FileIcon.tsx`, `ProviderIcon.tsx`.

Nothing else changes. Imports are rewritten to keep the house style — a sibling
module through `./name`, anything else through the `@/` alias — and the three
hook docs plus DESIGN.md now point at the new paths. tsc resolves every import,
so the move is proven, not assumed.
