+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import { HashRouter, Outlet, Route, Routes } from "react-router-dom"
-import { SettingsProvider } from "@/lib/settings"
-import { ProjectsProvider } from "@/lib/projects"
+import { SettingsProvider } from "@/providers/settings"
+import { ProjectsProvider } from "@/providers/projects"
 import { ProjectTabs } from "@/components/tabs/ProjectTabs"
 import { SessionSidebar } from "@/components/sidebar/SessionSidebar"
 import { TerminalHost } from "@/components/TerminalHost"
