+++ frontend/src/components/AppUpdateGate.tsx
@@ -4,7 +4,7 @@ import { useMatch, useNavigate } from "react-router-dom"
 import { Button } from "@/components/ui/button"
 import { decideUpdateAction, UPDATE_DISMISSED_KEY, type UpdateAction } from "@/lib/app-update-gate"
 import { AppUpdate, System } from "@/lib/rpc"
-import { useProjects } from "@/lib/projects"
+import { useProjects } from "@/providers/projects"
 import { queuePaste } from "@/lib/paste-queue"
 import { registerUpdateChecker } from "@/lib/update-check"
 import { errorText } from "@/lib/utils"
