+++ frontend/src/components/CommandPalette.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo, useRef, useState } from "react"
 import { Dialog as DialogPrimitive } from "@base-ui/react/dialog"
 import { useNavigate } from "react-router-dom"
 import { CornerDownLeft, Folder, Search } from "lucide-react"
-import { useProjects } from "@/lib/projects"
-import { useSettings } from "@/lib/settings"
+import { useProjects } from "@/providers/projects"
+import { useSettings } from "@/providers/settings"
 import { isRecordingTarget, matchesCombo } from "@/lib/hotkeys"
 import { useSessionStatus } from "@/lib/useSessionStatus"
 import { SessionStatusIcon } from "@/components/sidebar/SessionStatusIcon"
