+++ frontend/src/components/FooterBar.tsx
@@ -11,11 +11,11 @@ import { useSessionUsage } from "@/lib/useSessionUsage"
 import { displayPath } from "@/lib/paths"
 import { useGitStatus } from "@/lib/useGitStatus"
 import { usePullRequest } from "@/lib/usePullRequest"
-import { useSettings } from "@/lib/settings"
+import { useSettings } from "@/providers/settings"
 import { cn } from "@/lib/utils"
-import { DiffStat } from "@/components/DiffStat"
-import { ContextRing, contextColor } from "@/components/ContextRing"
-import { SessionModel } from "@/components/SessionModel"
+import { DiffStat } from "./DiffStat"
+import { ContextRing, contextColor } from "./ContextRing"
+import { SessionModel } from "./SessionModel"
 import { Separator } from "@/components/ui/separator"
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip"
