+++ frontend/src/components/Home.tsx
@@ -1,7 +1,7 @@
 import { FolderOpen } from "lucide-react"
 import { EmptyScreen } from "@/components/common/EmptyScreen"
 import { Button } from "@/components/ui/button"
-import { useProjects } from "@/lib/projects"
+import { useProjects } from "@/providers/projects"
 
 // Home is the landing screen shown when no project is active. It sits on top of
 // the (currently empty) terminal host and offers the OS directory picker.
