+++ frontend/src/components/TerminalView.tsx
@@ -18,8 +18,8 @@ import { takeSetup } from "@/lib/setup-queue"
 import { recordChunk } from "@/lib/term-perf"
 import { copyToastMessage, COPY_TOAST_DURATION_MS } from "@/lib/copy-toast"
 import { computeGrid } from "@/lib/term-fit"
-import { useSettings } from "@/lib/settings"
-import type { ResolvedTheme } from "@/lib/settings"
+import { useSettings } from "@/providers/settings"
+import type { ResolvedTheme } from "@/providers/settings"
 import type { SessionKind } from "@/lib/sessions"
 import "@xterm/xterm/css/xterm.css"
