+++ frontend/src/components/dock/FilesPanel.tsx
@@ -6,7 +6,7 @@ import { FileTree } from "@/components/FileTree"
 import { formatLineRef, parseDiff, type DiffFile } from "@/lib/diff"
 import { buildTree, type TreeNode } from "@/lib/file-tree"
 import { queuePaste } from "@/lib/paste-queue"
-import { useProjects } from "@/lib/projects"
+import { useProjects } from "@/providers/projects"
 import { ProjectService, System } from "@/lib/rpc"
 import { useActiveSession } from "@/lib/useActiveSession"
 import { useGitStatus } from "@/lib/useGitStatus"
