+++ frontend/src/components/settings/HotkeysSettings.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react"
 import { RotateCcw } from "lucide-react"
-import { useSettings } from "@/lib/settings"
+import { useSettings } from "@/providers/settings"
 import {
   comboFromEvent,
   DEFAULT_HOTKEYS,
