+++ frontend/src/components/settings/ProviderBinSettings.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react"
 import { Store } from "@/lib/rpc"
-import { useProjects } from "@/lib/projects"
+import { useProjects } from "@/providers/projects"
 import { binKey } from "@/lib/providers-store"
-import { useSettings } from "@/lib/settings"
+import { useSettings } from "@/providers/settings"
 import { Input } from "@/components/ui/input"
 import { Switch } from "@/components/ui/switch"
 import { SettingBlock } from "./SettingBlock"
