+++ frontend/src/components/tabs/ProjectTabs.tsx
@@ -4,7 +4,7 @@ import { DndContext, closestCenter } from "@dnd-kit/core"
 import { SortableContext, horizontalListSortingStrategy } from "@dnd-kit/sortable"
 import { cn } from "@/lib/utils"
 import { Button } from "@/components/ui/button"
-import { useProjects } from "@/lib/projects"
+import { useProjects } from "@/providers/projects"
 import { sessionsOf } from "@/lib/sessions"
 import { openSettings } from "@/lib/settings-card-store"
 import { NotificationsButton } from "./NotificationsButton"
