+++ frontend/src/providers/settings.tsx
@@ -8,8 +8,8 @@ import {
   type Combo,
   type HotkeyId,
   type Hotkeys,
-} from "./hotkeys"
-import { zoomIntent } from "./zoom-keys"
+} from "@/lib/hotkeys"
+import { zoomIntent } from "@/lib/zoom-keys"
 
 const FONT_STORAGE_KEY = "lich.terminal.font"
 const TERMINAL_FONT_SIZE_STORAGE_KEY = "lich.terminal.fontSize"
